requirements.txt
fastapi
uvicorn
streamlit
pymupdf
mindee
firecrawl-py
openai
sentence-transformers
scikit-learn
plotly
python-multipart
python-dotenv